tests/submission/Infraware/Forms/contents/Forms/progress_max.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>progress_max</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form onsubmit="return false" name='input_form'>
		 <p>Progress: <progress max=100 id='progress_id'></progress></p>
	</form>

	<script>
		
		var progress = document.getElementById("progress_id");
		progress.value = 50;
		var max = progress.max;
		progress.max = 200;

		if (typeof(progress.max) == "number") {
			if (max == 100 && progress.max == 200) {
				test(function() {
					assert_true(true, "max attribute is correct.");
				});
			} else {
				test(function() {
					assert_unreached("max attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("max attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
